% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_statistic.R
\name{random_statistic}
\alias{random_statistic}
\title{Random Statistics Function}
\usage{
random_statistic(n)
}
\description{
set n to any integer
n is for any random integer between 1 and that integer
will print the list of integers, sum, range and mean
}
