% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_of_consecutive_int.R
\name{sum_of_consecutive_int}
\alias{sum_of_consecutive_int}
\title{Sum of Consecutive Integers Function}
\usage{
sum_of_consecutive_int(n)
}
\description{
set n to any integer by first entering: n <- \emph{any integer you like}
will return the sum of consecutive integers of n
}
